Refactor viewer2d pure panel helpers into dedicated module
